ENH: Generate scripts for running registration instead of generating scripts in a subprocess; ENH: Refactor dataset classes into single file (avoiding from/import inconsistencies between Python 2.7 and 3.x): ENH: Allow blueprints to select specific datasets instead of categories
