+++ SuperBench/Source/evaluate.py
@@ -1,5 +1,5 @@
 import logging, os, json
-from register import load_submissions, load_datasets, get_category
+from make_registration_scripts import parser, load_submissions, load_datasets
 
 def write_json(output_file_name, results):
     output_directory = os.path.dirname(output_file_name)
@@ -15,20 +15,27 @@ def run(parameters):
 
     for team_name, blueprint_file_name in submissions:
         blueprint_name, blueprint_ext = os.path.splitext(os.path.basename(blueprint_file_name))
-        category = get_category(blueprint_file_name)
-        results = { team_name: { blueprint_name: dict() } }
-        for dataset in datasets[category]:
-            results[team_name][blueprint_name][dataset.name] = dict()
+
+        blueprint = json.load(open(blueprint_file_name))
+
+        results = {team_name: {blueprint_name: dict()}}
+        for dataset_name in blueprint['Datasets']:
+            if not dataset_name in datasets:
+                logging.error('Dataset ' + dataset_name + ' requested by ' + blueprint_file_name + ' but no directory provided. See \'--help\' for usage.')
+                continue
+
+            dataset = datasets[dataset_name]
+            results[team_name][blueprint_name][dataset_name] = dict()
             for file_names in dataset.generator():
                 output_directory = os.path.join(parameters.output_directory, team_name, blueprint_name)
                 logging.info('Evalualting %s and %s.', file_names.image_file_names[0], file_names.image_file_names[1])
+
                 try:
                     file_names.deformation_field_file_names_full_path = [
                         (os.path.join(output_directory, file_names.deformation_field_file_names[0]),
                          os.path.join(output_directory, file_names.deformation_field_file_names[1]))]
 
                     result_0, result_1 = dataset.evaluate(parameters.registration_driver, file_names)
-
                     results[team_name][blueprint_name][dataset.name][file_names.deformation_field_file_names[0]] = result_0
                     results[team_name][blueprint_name][dataset.name][file_names.deformation_field_file_names[1]] = result_1
                 except Exception as e:
@@ -37,3 +44,7 @@ def run(parameters):
                     logging.error('Error during evaulation of %s, %s, %s' % (team_name, blueprint_name, dataset.name))
 
         write_json(os.path.join(parameters.output_directory, "results.json"), results)
+
+if __name__ == '__main__':
+    parameters = parser.parse_args()
+    run(parameters)
