+++ SuperBench/Submissions/TeamStefan/elastix_lung.json
@@ -1,5 +1,5 @@
 {
-    "SuperBenchCategory": "Lung",
+    "Datasets": ["POPI"],
     "Component": {
         "Name": "Elastix",
         "NameOfClass": "MonolithicElastixComponent",
